JMO-23 Remove save button in SettingsWindow
